+++ packages/meteor-type-validation/tests/Humanize.test.ts
@@ -1,5 +1,5 @@
 import { expect, it } from 'vitest';
-import { humanizeProperty } from '../src/server/util/Humanize';
+import { humanizeProperty } from '../src/server/utils/Humanize';
 
 it('should split camelCase', () => {
     expect(humanizeProperty('createdBy')).toBe('Created by');
